used post processing, bling-phong light model, and reflections
For the post processing I used both ambient occlusion and chromatic aberration. In my opinion the model looks better with just the ambient occlusion added.

For the lighting model I used bling-phong lighting model with 3 sources of light form different points. and messed around with the colors.

There is actually also a texture added of the lava and I noticed At some point it was visible if some part of the light calculation was commented out or altered a bit then it would be visible maybe.

The codes for the reflection are in yes but the plane object for some reason is not visible in the screen as it may be out of sight or it may not have been rendered at all. But I did try to add it.

ps.- I fell sick towards the end so my efficiency decreased alot towards the end. I can't say I didn't have enough time but I did what I was able to do with the time I had.  And I cant speak, so I typed a simple understanding of what my codes are about and what I know I did right and where I messed up.